[hotfix][table-planner][table-api-java] Move QueryOperation factories to table-api-java
